[Java] Add Aeron benchmark scripts and config files (#5)
* [Java] Change default IdleStrategy to NoOpIdleStrategy

* [Java] Add basic scripts for Aeron benchmarks

* [Java] Add a new line before printing benchmark startup message

* [Java] Simplify ArchiveMessageTransceiver

* [Java] Spin on `startReplay`, i.e. await the creation of the segment file for the recording

* [Java] Naming

* [Java] Add scripts for all scenarios

* [Java] Fix NPE when closing nodes

* [Java] Fix path to the aeronmd binary

* [Java] Close resources even when Archive was gone

* [Java] Warn when the target rate was not achieved

* [Java] Correct scripts startup order

* [Java] Remove MTU config values and specify directory locations

* [Java] Use setMemory to fill up the rest of message when bigger than 8 bytes

* [Java] Add test for `burstSize > 1`

* [Java] Ensure proper cleanup up exception

* [Java] Check for interrupt

* [Java] Reduce number of messages and increase timeouts

* [Java] Reduce big message to 128K

* [Java] Remove creation debug

* [Java] Reduce big message size to 8K

* [Java] Reduce big message size to 8K

* [Java] Use 32K for big messages

* [Java] Add test with 200 byte messages

* [Java] Reduce number of messages

+++ src/main/java/uk/co/real_logic/benchmarks/rtt/aeron/EchoNode.java
@@ -74,7 +74,10 @@ public final class EchoNode implements AutoCloseable, Runnable
         if (ownsAeronClient)
         {
             closeAll(aeron, mediaDriver);
-            mediaDriver.context().deleteAeronDirectory();
+            if (null != mediaDriver)
+            {
+                mediaDriver.context().deleteAeronDirectory();
+            }
         }
     }
