+++ src/main/java/uk/co/real_logic/benchmarks/rtt/aeron/LiveReplayMessageTransceiver.java
@@ -26,7 +26,6 @@ import uk.co.real_logic.benchmarks.rtt.MessageTransceiver;
 
 import static io.aeron.ChannelUri.addSessionId;
 import static io.aeron.archive.client.AeronArchive.connect;
-import static java.lang.Long.MAX_VALUE;
 import static java.nio.ByteOrder.LITTLE_ENDIAN;
 import static org.agrona.BitUtil.CACHE_LINE_LENGTH;
 import static org.agrona.BufferUtil.allocateDirectAligned;
@@ -103,7 +102,10 @@ public final class LiveReplayMessageTransceiver extends MessageTransceiver
         if (ownsArchiveClient)
         {
             closeAll(aeronArchive, mediaDriver);
-            mediaDriver.context().deleteAeronDirectory();
+            if (null != mediaDriver)
+            {
+                mediaDriver.context().deleteAeronDirectory();
+            }
         }
     }
 
@@ -137,7 +139,7 @@ public final class LiveReplayMessageTransceiver extends MessageTransceiver
 
         final String replayChannel = receiveChannel();
         final int replayStreamId = receiveStreamId();
-        replaySessionId = aeronArchive.startReplay(recordingId, 0, MAX_VALUE, replayChannel, replayStreamId);
+        replaySessionId = replayFullRecording(aeronArchive, recordingId, replayChannel, replayStreamId);
 
         final String channel = addSessionId(replayChannel, (int)replaySessionId);
         subscription = aeronArchive.context().aeron().addSubscription(channel, replayStreamId);
